// src/app/shared/sdk/rest-api/model/jsonIocResponse.ts
/**
 * CyberReportHub API
 * The API to interact with CyberReportHub
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface JsonIocResponse {
  iocId: number;
  iocTypeId: number;
  iocTypeName: string;
  value: string;
}